# src/aic-website/api-docs/backend/openapi.yaml
openapi: 3.0.3
info:
  title: AIC Website Backend API
  description: |
    API documentation for the AIC Website Backend Services.
    
    This API provides access to core backend functionality including user management,
    content management, and application settings.
  version: 1.0.0
  contact:
    name: AIC Development Team
    email: dev@example.com
    url: https://developer.example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.example.com/v1
    description: Production API Server
  - url: https://staging-api.example.com/v1
    description: Staging API Server
  - url: https://dev-api.example.com/v1
    description: Development API Server
...
paths:
...
  /content:
    get:
      summary: Get all content
      description: Returns a list of content items
      operationId: getContent
      tags:
        - Content
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: type
          in: query
          description: Filter content by type
          required: false
          schema:
            type: string
            enum: [article, page, product]
        - name: status
          in: query
          description: Filter content by status
          required: false
          schema:
            type: string
            enum: [draft, published, archived]
      responses:
        '200':
          description: List of content items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create content
      description: Creates a new content item
      operationId: createContent
      tags:
        - Content
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentRequest'
      responses:
        '201':
          description: Content created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: Sort field and direction (e.g., name:asc, createdAt:desc)
      required: false
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
        path:
          type: string
      required:
        - error
        - message
        - statusCode
        - timestamp
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, editor, user]
        status:
          type: string
          enum: [active, inactive, pending]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - status
        - createdAt
        - updatedAt
...
    Content:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [article, page, product]
        status:
          type: string
          enum: [draft, published, archived]
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
        author:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - type
        - status
        - content
        - author
        - createdAt
        - updatedAt
    ContentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Content'
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
        - data
        - pagination
    ContentResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Content'
      required:
        - data
    CreateContentRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [article, page, product]
        status:
          type: string
          enum: [draft, published, archived]
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
      required:
        - title
        - type
        - content
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - totalItems
        - totalPages